refactor: split preflight probe and probe-result modules under size and drift caps
